Convert flag array from int32 to uint8 in v2.
Try to save on some memory and IO transfer costs.

+++ montblanc/impl/biro/v2/gpu/RimeGaussBSum.py
@@ -44,6 +44,7 @@ DOUBLE_PARAMS = {
 
 KERNEL_TEMPLATE = string.Template("""
 #include <cstdio>
+#include <stdint.h>
 #include \"math_constants.h\"
 #include <montblanc/include/abstraction.cuh>
 
@@ -77,7 +78,7 @@ void rime_gauss_B_sum_impl(
     typename Tr::ft * wavelength,
     int * ant_pairs,
     typename Tr::ct * jones_EK_scalar,
-    int * flag,
+    uint8_t * flag,
     typename Tr::ft * weight_vector,
     typename Tr::ct * visibilities,
     typename Tr::ct * data_vis,
@@ -406,7 +407,7 @@ rime_gauss_B_sum_ ## symbol ## chi_ ## ft( \
     ft * wavelength, \
     int * ant_pairs, \
     ct * jones_EK_scalar, \
-    int * flag, \
+    uint8_t * flag, \
     ft * weight_vector, \
     ct * visibilities, \
     ct * data_vis, \
